+++ tests/core/test_views_errors.py
@@ -347,6 +347,11 @@ class TrailingSlashRedirectViewTest(FormattedErrorMixin, BaseWebTest, unittest.T
         self.assertEqual(response.status_int, 307)
         self.assertEqual(response.location, "http://localhost/v0/")
 
+    def test_does_contain_csp_header(self):
+        response = self.app.get("")
+        self.assertEqual(response.status_int, 307)
+        self.assertIn("default-src 'none'", response.headers["Content-Security-Policy"])
+
     def test_it_redirects_if_it_ends_with_a__slash_(self):
         response = self.app.get("/mushrooms/")
         self.assertEqual(response.status_int, 307)